Remove footer from allocated blocks
This modification intends to make better use of memory
by decreasing memory overhead due to the usage of
a header/footer in each allocated block.

Footer are removed from allocated blocks and contain only
a header. Free blocks still contain a header and a footer.

As a consequence, the minimum block size decreases from
16 bytes to 8 bytes, and a block only "knows" the size of
its previous block if this previous block is currently free.