Implement a mixed version of segregated lists
One of the biggest problems of the first implementation
of segregated list algorithm is that it takes too much
processing time and space to allocated small blocks
(which are the most common in most applications).

The current implementation tries to address this problem
by associating fixed-size size classes with all block sizes
lower than a certain amount (2kB in the current implementation).
It decreases a significant part of the overhead to allocate
small blocks. We still use mixed-size size classes for larger
classes, though.

As a consequence, the new minimum block size decreased from
40 to 24.